remove dom/types.h
